+++ js/massive.js
@@ -1,0 +1,54 @@
+const DESCRIPTION = [
+    'Бухта у отеля',
+    'Знак',
+    'Пляж на тропическом острове',
+    'Чика',
+    'Довольный рис',
+    'Клевая тачка',
+    'Клубника',
+    'Сок из смородины',
+    'Чика и самолет',
+    'Обувница',
+    'Дорога к пляжу',
+    'Ауди',
+    'Вегетарианское блюдо',
+    'Суши кот',
+    'Дурное воспитание',
+    'Вид из окна самолета',
+    'Оркестр',
+    'Раритет',
+    'Тапки с фонариком',
+    'Вид на отель',
+    'Салат',
+    'Закат на пляже',
+    'Краб',
+    'Нашествие',
+    'Сафари',
+  ];
+
+  const MESSAGE = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  ];
+
+  const NAMES = [
+    'Иван',
+    'Алексей',
+    'Дмитрий',
+    'Вячеслав',
+    'Михаил',
+    'Роман',
+    'Виталий',
+    'Андрей',
+    'Антон',
+    'Станислав',
+    'Богдан',
+    'Руслан',
+    'Юрий',
+  ];
+
+  export {DESCRIPTION, MESSAGE, NAMES};
